rename common code as Settings.ts
